+++ webapp/loggedin/appHeader/appHeader.js
@@ -6,7 +6,10 @@ import { useSelector, useDispatch } from 'react-redux'
 import * as User from '@core/user/user'
 import * as Survey from '@core/survey/survey'
 
-import { usePrevious, useUser, useSurveyCycleKey, useSurveyInfo } from '@webapp/components/hooks'
+import { useLang } from '@webapp/store/system'
+import { useSurveyCycleKey, useSurveyInfo } from '@webapp/store/survey'
+
+import { usePrevious, useUser } from '@webapp/components/hooks'
 import ProfilePicture from '@webapp/components/profilePicture'
 import ProgressBar from '@webapp/components/progressBar'
 import ButtonPublishSurvey from '@webapp/components/buttonPublishSurvey'
@@ -15,7 +18,6 @@ import * as AppState from '@webapp/app/appState'
 import { updateUserPrefs } from '@webapp/app/actions'
 import UserPopupMenu from './components/userPopupMenu'
 import CycleSelector from './components/cycleSelector'
-import { useLang } from '@webapp/store/system'
 
 const AppHeader = () => {
   const dispatch = useDispatch()
