+++ webapp/loggedin/modules/analysis/calculation/view/useCalculationState.js
@@ -8,8 +8,8 @@ import * as Chain from '@common/analysis/processingChain'
 import * as Step from '@common/analysis/processingStep'
 import * as Calculation from '@common/analysis/processingStepCalculation'
 
-import { useNodeDefByUuid, useSurvey, useSurveyInfo } from '@webapp/components/hooks'
 import { useI18n } from '@webapp/store/system'
+import { useSurvey, useSurveyInfo, useNodeDefByUuid } from '@webapp/store/survey'
 import { useChainEdit } from '@webapp/loggedin/modules/analysis/hooks'
 
 import * as ChainState from '@webapp/loggedin/modules/analysis/chain/state'
