+++ webapp/loggedin/modules/analysis/chains/view/chainsView.js
@@ -5,7 +5,8 @@ import { useHistory } from 'react-router'
 
 import * as Chain from '@common/analysis/processingChain'
 
-import { useSurveyCycleKey } from '@webapp/components/hooks'
+import { useSurveyCycleKey } from '@webapp/store/survey'
+
 import TableView from '@webapp/loggedin/tableViews/tableView'
 
 import { navigateToChainView } from '@webapp/loggedin/modules/analysis/chains/actions'
