+++ webapp/loggedin/modules/data/dataVis/dataQuery/components/useTableState.js
@@ -4,17 +4,17 @@ import * as R from 'ramda'
 import { ColumnNodeDef } from '@common/model/db'
 import * as Survey from '@core/survey/survey'
 import { WebSocketEvents } from '@common/webSocket/webSocketEvents'
-
 import * as AppWebSocket from '@webapp/app/appWebSocket'
 
-import { useOnUpdate, useSurvey, useAuthCanCleanseRecords } from '@webapp/components/hooks'
+import { useLang } from '@webapp/store/system'
+import { useSurvey } from '@webapp/store/survey'
 
+import { useOnUpdate, useAuthCanCleanseRecords } from '@webapp/components/hooks'
 import * as AppState from '@webapp/app/appState'
 import * as NodeDefUIProps from '@webapp/loggedin/surveyViews/surveyForm/nodeDefs/nodeDefUIProps'
 import { nodesUpdateCompleted, nodeValidationsUpdate } from '../actions'
 
 import * as DataQueryState from '../state'
-import { useLang } from '@webapp/store/system'
 
 export const useTableState = () => {
   const dispatch = useDispatch()
