+++ webapp/loggedin/modules/designer/surveyHierarchy/surveyHierarchyView.js
@@ -5,8 +5,9 @@ import React, { useEffect, useState, useRef } from 'react'
 import * as Survey from '@core/survey/survey'
 import * as NodeDef from '@core/survey/nodeDef'
 
-import { useSurvey } from '@webapp/components/hooks'
 import { useI18n } from '@webapp/store/system'
+import { useSurvey } from '@webapp/store/survey'
+
 import NodeDefsSelectorView from '../../../surveyViews/nodeDefsSelector/nodeDefsSelectorView'
 import Tree from './surveyHierarchyTree'
