+++ webapp/loggedin/modules/home/collectImportReport/components/tableRow.js
@@ -6,10 +6,11 @@ import PropTypes from 'prop-types'
 import * as Survey from '@core/survey/survey'
 import * as NodeDef from '@core/survey/nodeDef'
 import * as CollectImportReportItem from '@core/survey/collectImportReportItem'
-
 import { appModuleUri, designerModules } from '@webapp/app/appModules'
-import { useSurvey } from '@webapp/components/hooks'
+
 import { useI18n } from '@webapp/store/system'
+import { useSurvey } from '@webapp/store/survey'
+
 import Checkbox from '@webapp/components/form/checkbox'
 import LabelsEditor from '@webapp/loggedin/surveyViews/labelsEditor/labelsEditor'
