+++ webapp/loggedin/surveyViews/nodeDefsSelector/components/attributesSelector.js
@@ -5,9 +5,9 @@ import * as PropTypes from 'prop-types'
 import * as Survey from '@core/survey/survey'
 import * as NodeDef from '@core/survey/nodeDef'
 
-import { useSurvey } from '@webapp/components/hooks'
-import ExpansionPanel from '@webapp/components/expansionPanel'
+import { useSurvey } from '@webapp/store/survey'
 
+import ExpansionPanel from '@webapp/components/expansionPanel'
 import AttributeSelector from './attributeSelector'
 
 const AttributesSelector = (props) => {
