+++ webapp/store/survey/index.js
@@ -18,4 +18,14 @@ export { CategoriesActions } from './categories'
 export { TaxonomiesActions } from './taxonomies'
 
 // ====== hooks
-export { useSurveyDefsFetched } from './hooks'
+export {
+  useSurveyDefsFetched,
+  useSurvey,
+  useSurveyId,
+  useSurveyInfo,
+  useSurveyCycleKey,
+  useSurveyLang,
+  useOnSurveyCycleUpdate,
+  useNodeDefByUuid,
+  useCategoryByUuid,
+} from './hooks'
